# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Laravel",
        "description": "Implémentation de Swagger avec L5-Swagger.",
        "contact": {
            "name": "KANHONOU AUSCENCE",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de l'API"
        }
    ],
    "paths": {
...
        "/api/organisme/store": {
            "post": {
                "tags": [
                    "Organisme"
                ],
                "summary": "Crée un nouvel Organisme",
                "operationId": "9fbab55e02324a9f50b585e2ec1ee347",
                "requestBody": {
                    "description": "Organisme à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Organisme"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Si( Success = true) Organisme créé avec succès   ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organisme"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Les données envoyées sont incorrectes"
                    }
                }
            }
        },
        "/api/organisme/show/{organisme}": {
            "get": {
                "tags": [
                    "Organisme"
                ],
                "summary": "Récupère un organisme",
                "description": "Retourne l'organisme identifié par son ID.",
                "operationId": "f6d0219032920d0e8e48cbf4993ecc48",
                "parameters": [
                    {
                        "name": "organisme",
                        "in": "path",
                        "description": "Identifiant de l'organisme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organisme retourné",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organisme"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organisme non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organisme/update/{organisme}": {
            "put": {
                "tags": [
                    "Organisme"
                ],
                "summary": "Met à jour un organisme existant",
                "description": "Met à jour un organisme identifié par son ID.",
                "operationId": "updateOrganisme",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/organisme--id"
                    }
                ],
                "requestBody": {
                    "description": "Données de l'organisme à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Organisme"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organisme mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organisme"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Organisme non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organisme/delete/{id}": {
            "delete": {
                "tags": [
                    "Organisme"
                ],
                "summary": "Supprime un Organisme",
                "operationId": "c982019280eb19ed858437aeda5d4cd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l' Organisme à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Si (Success = true) Organisme supprimé avec succès "
                    },
                    "404": {
                        "description": "Organisme non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Organisme": {
                "title": "Organisme",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifiant de l'Organisme",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Nom de l'Organisme",
                        "type": "string",
                        "example": "Présidence"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "organisme--id": {
                "name": "organisme_id",
                "in": "path",
                "description": "Identifiant de organisme",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}